Simplify calls to deezerApp so that all initializations are encapsulated
